pyproject.toml
[project]
name = "phase-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=9.4.0",
    "pytest>=8.4.2",
]
